requirements.txt
argcomplete==3.5.1
argh==0.31.3
asciitree==0.3.3
biopython==1.83
blosc2==2.0.0
bx-python==0.13.0
click==8.1.7
contourpy==1.1.1
cooler==0.10.2
cycler==0.12.1
Cython==3.0.11
cytoolz==1.0.0
dill==0.3.9
exceptiongroup==1.2.2
fonttools==4.54.1
future==1.0.0
gffutils==0.13
h5py==3.11.0
HiCMatrix==17.2
HTSeq==2.0.9
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
intervaltree==3.1.0
kiwisolver==1.4.7
matplotlib==3.7.5
msgpack==1.1.0
multiprocess==0.70.17
natsort==8.4.0
numexpr==2.8.6
numpy==1.24.4
packaging==24.1
pandas==2.0.3
pillow==10.4.0
plotly==5.24.1
pluggy==1.5.0
py-cpuinfo==9.0.0
pybedtools==0.10.0
pyBigWig==0.3.22
pyfaidx==0.8.1.3
pyGenomeTracks==3.9
pyparsing==3.1.4
pysam==0.22.1
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
scipy==1.10.1
simplejson==3.19.3
six==1.16.0
sortedcontainers==2.4.0
tables==3.8.0
tenacity==9.0.0
tomli==2.0.2
toolz==1.0.0
tqdm==4.66.5
tzdata==2024.2
zipp==3.20.2
